+++ examples/streaming-tour/src/client.rs
@@ -45,7 +45,7 @@ async fn main() -> Result<(), BoxError> {
         if !first {
             print!(", ");
         }
-        print!("{}", msg.reborrow().value.unwrap_or_default());
+        print!("{}", msg.view().value.unwrap_or_default());
         first = false;
     }
     println!("]");
@@ -79,7 +79,7 @@ async fn main() -> Result<(), BoxError> {
             if !first {
                 print!(", ");
             }
-            print!("{}", msg.reborrow().total.unwrap_or_default());
+            print!("{}", msg.view().total.unwrap_or_default());
             first = false;
         }
     }